playlists page small change
